+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Version 7.2
 
+### Version 7.2.0.3 - 2024-03-13
+
+#### Added
+-   Support for multiple column attribute overrides in `GPUdbTable.from_df()`
+
+#### Changed
+-   Increased connection timeout from 1 to 20 seconds to account for
+    connections over high-traffic and public networks
+
+
 ### Version 7.2.0.2 - 2024.03.06
 
 #### Added
@@ -34,6 +44,16 @@
 
 
 ## Version 7.1
+
+### Version 7.1.9.11 - 2024-03-13
+
+#### Added
+-   Support for multiple column attribute overrides in `GPUdbTable.from_df()`
+
+#### Changed
+-   Increased connection timeout from 1 to 20 seconds to account for
+    connections over high-traffic and public networks
+
 
 ### Version 7.1.9.10 - 2023-11-13